ContextMenu: add runtime options (#143)
